Fix decision making if SSID is human readable
It was simply broken. It it supposed to mark all networks as NOT human readable if their SSID only contains of whitespaces or control characters. However, it marked them as NOT human readable if they contained at least one whitespace or control character and, by that, marking perfeclty fine networks as NOT human readable.